Mirror integration test files to the level modules
Rename tests/validate.rs -> validate_schema.rs (dropping the now-redundant
schema_ prefix on its test fns and snapshots), and split tests/compare.rs into
validate_meta.rs (names + types) and validate_data.rs (value checks), with the
shared parquet/dict fixtures in tests/common/mod.rs. Each test file now exercises
its level's entry point directly.
