+++ crates/data-dict/tests/common/mod.rs
@@ -1,0 +1,73 @@
+//! Shared fixtures for the metadata and data comparison integration tests.
+//!
+//! Each helper writes a small parquet file and/or dictionary YAML to a unique
+//! temp dir. Both `validate_meta` and `validate_data` test binaries include this
+//! module, so not every helper is used by each — hence the blanket allow.
+#![allow(dead_code)]
+
+use std::fs::File;
+use std::path::{Path, PathBuf};
+use std::sync::Arc;
+use std::sync::atomic::{AtomicU32, Ordering};
+
+use indoc::indoc;
+use parquet::data_type::{ByteArray, ByteArrayType, DoubleType};
+use parquet::file::properties::WriterProperties;
+use parquet::file::writer::SerializedFileWriter;
+use parquet::schema::parser::parse_message_type;
+
+static COUNTER: AtomicU32 = AtomicU32::new(0);
+
+/// A unique temp directory for one test's fixtures.
+pub fn temp_dir() -> PathBuf {
+    let mut dir = std::env::temp_dir();
+    dir.push(format!(
+        "data-dict-test-{}-{}",
+        std::process::id(),
+        COUNTER.fetch_add(1, Ordering::Relaxed)
+    ));
+    std::fs::create_dir_all(&dir).unwrap();
+    dir
+}
+
+/// Write a parquet file with a string `name` column and a double `weight`
+/// column.
+pub fn write_parquet(path: &Path) {
+    let message = indoc! {"
+        message schema {
+            REQUIRED BYTE_ARRAY name (UTF8);
+            REQUIRED DOUBLE weight;
+        }
+    "};
+    let schema = Arc::new(parse_message_type(message).unwrap());
+    let props = Arc::new(WriterProperties::builder().build());
+    let file = File::create(path).unwrap();
+    let mut writer = SerializedFileWriter::new(file, schema, props).unwrap();
+    let mut rg = writer.next_row_group().unwrap();
+
+    let mut col = rg.next_column().unwrap().unwrap();
+    col.typed::<ByteArrayType>()
+        .write_batch(
+            &[ByteArray::from("otter"), ByteArray::from("seal")],
+            None,
+            None,
+        )
+        .unwrap();
+    col.close().unwrap();
+
+    let mut col = rg.next_column().unwrap().unwrap();
+    col.typed::<DoubleType>()
+        .write_batch(&[1.0_f64, 2.0], None, None)
+        .unwrap();
+    col.close().unwrap();
+
+    rg.close().unwrap();
+    writer.close().unwrap();
+}
+
+/// Write `yaml` to `<dir>/dict.yaml` and return the path.
+pub fn write_yaml(dir: &Path, yaml: &str) -> PathBuf {
+    let path = dir.join("dict.yaml");
+    std::fs::write(&path, yaml).unwrap();
+    path
+}
